requirements.txt
flask
requests
redis
pycryptodome
